Okay, prerasterized font information is populated. Now josh just needs to do something with it.
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@502 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ pluginsource/org/enigma/backend/sub/Glyph.java
@@ -6,9 +6,9 @@ import com.sun.jna.Structure;
 
 public class Glyph extends Structure
 	{
-	public int origin;
-	public int baseline;
-	public int advance;
+	public double origin;
+	public double baseline;
+	public double advance;
 	public int width;
 	public int height;
 	public Buffer raster; //size = width * height
